+++ docs/CHANGES.md
@@ -13,6 +13,7 @@
 - #573 xgettext build error: `'javascript' unknown` [jcbrand]
 - #587 Fix issue when logging out with `auto_logout=true` [davec82]
 - #589 Save scroll position on minimize and restore it on maximize [rlanvin]
+- #592 Add random resource for auto_login, add method generateResource to generate random resource [davec82]
 
 ## 0.10.1 (2016-02-06)